some quality of life , still working so not a full changes log
