+++ excise_app/urls.py
@@ -1,10 +1,12 @@
 # excise/urls.py
-from django.urls import path
+from django.urls import path , include
 from .views import *
+
+
 
 urlpatterns = [
     # User Registration/Signup
-    path('register/', UserRegistrationView.as_view(), name='user-register'),
+    path('',include('user.urls') ),
 
     #getcaptcha endpoint
     # sends the captcha to user 
@@ -14,15 +16,15 @@ urlpatterns = [
     # Login using Username & Password
     # authenticates the user and checks for the captcha
 
-    path('login/', UserLoginView.as_view(), name='user-login'),
+    # path('login/', UserLoginView.as_view(), name='user-login'),
     
     # Login using Phone Number & OTP
     path('send_otp/', SendOTP.as_view(), name='send-otp'),
     path('otp_login/', OTPLoginView.as_view(), name='otp-login'),
 
     #UserDetails endpoint
-    path('userdetails/',UserDetails.as_view(), name='user_details'),
-    path('users/', UserListView.as_view(), name='user-list'),
+    # path('userdetails/',UserDetails.as_view(), name='user_details'),
+    # path('users/', UserListView.as_view(), name='user-list'),
 
     #DashboardCount View
     path('dashboard/', DashboardCountView.as_view(), name='dashboard'),
